Add transactional annotation
