+++ src/main/java/com/markruler/aptzip/service/AuthService.java
@@ -5,8 +5,10 @@ import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-@Service("authService")
+@Transactional
+@Service
 public class AuthService {
 
   private JavaMailSender javaMailSender;
