+++ src/main/java/com/markruler/aptzip/service/BoardService.java
@@ -28,9 +28,10 @@ import org.springframework.ui.Model;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-@Service
 @Slf4j
 @RequiredArgsConstructor
+@Transactional
+@Service
 public class BoardService {
   private final BoardRepository boardRepository;
   private final LikeRepository likeRepository;
