+++ src/main/java/com/markruler/aptzip/service/ConfirmationService.java
@@ -5,11 +5,13 @@ import com.markruler.aptzip.domain.user.UserAccountEntity;
 import com.markruler.aptzip.persistence.user.ConfirmationTokenRepository;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 
-@Service
 @RequiredArgsConstructor
+@Transactional
+@Service
 public class ConfirmationService {
   private final ConfirmationTokenRepository confirmationTokenRepository;
